Bug: Invocation of hashCode on array
in org.eclipse.help.internal.workingset.AdaptableTocsArray.hashCode()

The code invokes hashCode on an array. Calling hashCode on an array
returns the same value as System.identityHashCode, and ignores the
contents and length of the array. If you need a hashCode that depends on
the contents of an array a, use java.util.Arrays.hashCode(a).

Rank: Scary (5), confidence: Normal
Pattern: DMI_INVOKING_HASHCODE_ON_ARRAY
Type: DMI, Category: CORRECTNESS (Correctness)

Change-Id: If95318b7a877d96d070e17bfc5904d263b18df2e
Signed-off-by: Carsten Hammer <carsten.hammer@t-online.de>