Bug 569641 - LineNumberRulerColumn sometimes wrong
Prior to this patch, the LineNumberRulerColumn was only looking at the
"bounds" of the visible area to decide whether to redraw lines or not.
We now look at the line position and size as there can be some changes
(code mining added and removed) which do not affect bounds but should
affect line numbers rules.

Change-Id: I913b6ccaefb7ba78c65daef1d85f9294b9163d92
Signed-off-by: Mickael Istria <mistria@redhat.com>