Bug 568991 - Fix black headers in forms in system theme
CSS definitions for 'color' and 'text-background-color' had foreground
and background mixed up. Just swapping the values gives on OS X 10.14.6
with the default system "Accent color" set to "Blue" a form header with
black text on dark blue background, which is not easily readable. Use
the lighter #COLOR_TITLE-BACKGRUND-GRADIENT instead. That is the same
color as used by the section headers, though, and thus some distinction
is lost.

Includes version bump for 4.19.

Change-Id: I321d536531d1214efae43bdb655eb4853a06420b
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>