Bug 479218 - Add projects as resource to select in Open project Dialog
Users requested projects to be displayed in the Open
Resource Dialog. This allows for searching, filtering and selecting
projects in a quick and convenient way.

Changes:
OpenResourceHandler.java
- Added IProject as valid resource type for the OpenResourceDialog

FilteredResourcesSelectionDialog.java
- The OpenResourceDialog displays information about the parent of the
displayed resources. Since projects have usually ROOT as parent and
ROOT doesn't contain any useful information for the user, the
implementation for this case was changed to either omit the parent
information or use the selected resource instead.
- Added convenience method parentIsRoot(IResource) to check if resource
parent is of type ROOT.

OpenResourceDialog.java
- When Projects are selected, "Open" and "Open With" would not make
sense. Disabled this buttons for this use case.

Change-Id: I3f03f5aadff19b8498ace5d11a16df50604a8e5b
Signed-off-by: Dieter Mai <maidieter@googlemail.com>