Bug 571164: Two IContainer instances can share same location
When updating the container for a launch configuration, make sure that
the new and the old container do not point to the same location.
Nested projects can have different container instances that points to
the same location and a switch between these containers should not be
considered a "move" of the launch configuration.

Contributed by STMicroelectronics

Change-Id: I504f351b8ae9a6544c86659f99151a81059a7e38
Signed-off-by: Torbjörn Svensson <torbjorn.svensson@st.com>