Bug 552606 - Don't loose builder args for clean
When cleaning a C/C++ project using the headless build application, the
"active" build configuration will always be used and not the one
specified on the command line. To fix the issue, the clean method needs
to preserve the desired project/build configuration and not loose track
of it midway.

Contributed by STMicroelectronics

Also-by: Torbjörn Svensson <torbjorn.svensson@st.com>
Also-by: Jonah Graham <jonah@kichwacoders.com>
Signed-off-by: Torbjörn Svensson <torbjorn.svensson@st.com>
Signed-off-by: Jonah Graham <jonah@kichwacoders.com>
Change-Id: I9ffc78e557fbee8e3ab7dd640de30bfb1869e654