Bug 563497: Vertical menu bar handles are stretched
The handle image was stretched to the width of the toolbar if the
vertical toolbar is wider then the handle image. This can easily be seen
for the "File" toolbar. This toolbar contains an item that has
a big width (30px) as it has a drop down. The handle image is 22px
wide. This 22px was stretched to 30px.

The same is true for horizontal toolbars. If the handle icon's height is
smaller then the height of the toolbar it was stretched. We did not see
this up to now because the height of the handle was identical to the
height of the toolbar.

This change now no longer stretches the handle image and centers it
horizontally / vertical.

Change-Id: Ieb104c4b4672db2b165e6ccbfcc9cf1c766e3dca