Bug 571133: Fix sporadic test failures
The FileSystemResourceManagerTest resulted in throwing several
exceptions periodically on running but was not obvious what the cause
was.

One possibility is that many of the tests read and wrote the same file
name `foo` which on a slow network drive could have resulted in tests
colliding. Changing the files to be unique names per test seems to have
improved this condition, but the issue is by no means fixed for sure.

Change-Id: I481b6e7b1b083b6b2e2aa04effa4cbb2e9f3534f
Signed-off-by: Alex Blewitt <alex.blewitt@gmail.com>