Bug 572598 - Reset perspective destroys toolbar of shared part
When a perspective is reset, all old stacks are removed. This calls
removeGui on all shared elements. Shared part disconnect from the
placeholders in the stacks to be disposed, see ElementReferenceRenderer.
However, visible toolbars remain in the stacks, therefore the toolbar
are disposed. As a result, the toolbar is no longer visible in other
placeholders.

Before the gui is removed, hideChild is called on the parent renderer.
The LazyStackRenderer should be hiding the toolbar of the active part on
this call anyhow. (And only if it is currently showing the shared part.)
Side effect of this is that the toolbar is moved to the limbo shell, and
is no longer contained in the widgets of the stack. As a result, the
toolbar is also not disposed when the stack is disposed.

Change-Id: Ic1ad9b0d035fcf8262387a00bdd987c7258df925
Signed-off-by: Rolf Theunissen <rolf.theunissen@gmail.com>
Reviewed-on: https://git.eclipse.org/r/c/platform/eclipse.platform.ui/+/179050
Tested-by: Platform Bot <platform-bot@eclipse.org>