Bug 573221 - ListSelectionDialog: Improve usability and extend its API
Improve the usability of org.eclipse.ui.dialogs.ListSelectionDialog
(which is used e.g. in the "Save Resources" dialog: see bug 572939) and
extend its API in the following ways:

(a) Enabling/Disabling of the "Select All" and the "Deselect All"
buttons
Actual: Both buttons are always enabled.
Expected:
- When all elements are selected, the "Select All" button should be
disabled, otherwise enabled.
- When no elements are selected, the "Deselect All" button should be
disabled, otherwise enabled.

(b) Extend API
The class is marked as "@noextend This class is not intended to be
subclassed by clients.", but internally, it is extended several times
for different purposes. The goal is that subclassing should no longer be
necessary for customizations, to achieve more consistency and avoid
duplicate code, and to give clients the same capabilities as the dialogs
used internally.
The extended API should provide following additional possibilities via a
fluent API using the builder pattern:
- Dialog title (easier than via setTitle() after creating the dialog)
- OK button label:
  - When nothing is selected: For example, "Launch without Saving"
  - When one or more elements are selected: With {0} as placeholder for
the current selected elements and with {1} for the total number of
elements; for example, "Save {0} of {1} and Launch"
- Whether the dialog can be canceled (via Cancel button or by closing
the dialog window)
- Whether to use SWT#SHEET (modal dialog that is attached to a parent
window)
- Optional checkbox:
  - Label
  - Initial value
The existing constructor should be marked as deprecated in favor of the
new API.

Change-Id: I7a4f548bc41a346685751220e9853ee36df8e344
Signed-off-by: Holger Voormann <eclipse@voormann.de>
Reviewed-on: https://git.eclipse.org/r/c/platform/eclipse.platform.ui/+/179930
Tested-by: Platform Bot <platform-bot@eclipse.org>
Reviewed-by: Lars Vogel <Lars.Vogel@vogella.com>