Bug 573595 - AutoBuildJob following a search and replace might run as system job
Don't schedule autobuild job if autobuild is off.

This prevents all kinds of build job state inconsistencies shown in the
Progress view - if the autobuild is disabled, it is simply not running,
if it is enabled, it always is shown in the Progress.

With that, there is no need to set the "system" state for the autobuild
job to hide it if it runs but actually won't do anything.

Change-Id: Ib99c459822d8bc55654b46aee064ecefbbc3c2e1
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>
Reviewed-on: https://git.eclipse.org/r/c/platform/eclipse.platform.resources/+/180722
Tested-by: Platform Bot <platform-bot@eclipse.org>