Bug 568299 - changed default to allow early exit from inner loop
If one config requested rebuild, running following configs in same round
only wastes resources. We skip the rest of the current build round and
start again with the first config to be built, unless  a system property
is set to keep the legacy behavior.

The flag can be set via system property
-Dorg.eclipse.core.resources.disallowEarlyInnerBuildLoopExit and by
default early exit is allowed.

Change-Id: I763f51310240c786e443890c9aef6b205ca0076e
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>
Reviewed-on: https://git.eclipse.org/r/c/platform/eclipse.platform.resources/+/181910
Tested-by: Platform Bot <platform-bot@eclipse.org>