Bug 568299 - don't exit early build loop in "autobuild off" case
The "early exit" change is required only for autobuild case, where we
cancel the build because we know a rebuild is requested and next round
will be there. With autobuild off there is no next round, so that would
be a regression to exit early, therefore don't exit the build loop.

Change-Id: I4c059b4cbae3109a1b796cce18883c09d78748f7
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>
Reviewed-on: https://git.eclipse.org/r/c/platform/eclipse.platform.resources/+/181960
Tested-by: Platform Bot <platform-bot@eclipse.org>
Reviewed-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>