Bug 574409 - Forward getAdapter to wrapped object in E4PartWrapper
The E4PartWrapper does not consult the wrapped object for adaption.
Therefore, customization of e.g. the PropertiesView is not supported for
E4 parts.
To fix this, call adapt on the wrapped object. Also, still fall back to
the default case, because an adapter factory for E4PartWrapper could
have been registered in the platform.

Change-Id: I7df73db7943453ff57952b81996ea45c6b24a276
Signed-off-by: Rolf Theunissen <rolf.theunissen@gmail.com>
Reviewed-on: https://git.eclipse.org/r/c/platform/eclipse.platform.ui/+/183100