Bug 575453 - Doublecklick on Marker without File shows PackageExplorer
Since "Goto Resource" is disabled for markers without path and double
click on a Marker can not open any Editor:

Only if no File Editor could be opened this patch uses the first "Show
in" Handler that supports ISetSelectionTarget to select the Resource
(i.e. the Project).
In the Java workbench this defaults to the Java PackageExplorer.
(The ProjectExplorer does support it too.)

Note that the "Goto Resource" context menu is still disabled.

Change-Id: If719099d83a86376dcbea022f64cee5855679154
Signed-off-by: Joerg Kubitz <jkubitz-eclipse@gmx.de>
Reviewed-on: https://git.eclipse.org/r/c/platform/eclipse.platform.ui/+/184395
Tested-by: Platform Bot <platform-bot@eclipse.org>
Reviewed-by: Mickael Istria <mistria@redhat.com>