Bug 575893 - [performance] improve file search: avoid synchronization
FileSearchQuery did spend the most time in waiting for synchronization.
By using a ConcurrentHashMap the lock congestion can be avoided

Requires flushMatches AFTER searching a file (i.e of the single current
file of the thread) - not at start (i.e. of all other files which are
handled in parallel).

Change-Id: If412e738dcf3a36034836c9a2f0a2af1af839ddd
Signed-off-by: Joerg Kubitz <jkubitz-eclipse@gmx.de>
Reviewed-on: https://git.eclipse.org/r/c/platform/eclipse.platform.text/+/185196
Tested-by: Platform Bot <platform-bot@eclipse.org>