Bug 576169 - Distinguish between project removal from workspace/disk
On deleting a project from Package Explorer one can have two options:
(1) Remove from the workspace ("Delete project contents on disk (cannot
be undone)" checkbox is *not* selected)
(2) Remove from the workspace and disk ("Delete project contents on disk
(cannot be undone)" checkbox is selected)

This change adds a new flag to the IResourceDelta and a new method to
IResourceChangeDescriptionFactory to differentiate the
"type" of delete ((1) or (2)) in the delta objects created by
IResourceChangeDescriptionFactory.

ResourceChangeDescriptionFactory is updated to to set this flag
accordingly & tests added for new behavior.

Change-Id: Id3036344d864fb57fd1c514dbba3780df93adbbf
Signed-off-by: Mykola Zakharchuk <zakharchuk.vn@gmail.com>
Reviewed-on: https://git.eclipse.org/r/c/platform/eclipse.platform.resources/+/185701
Tested-by: Platform Bot <platform-bot@eclipse.org>
Reviewed-by: Andrey Loskutov <loskutov@gmx.de>