Bug 576333 - [performance] optimize EclipseContext.getChildren()
EclipseContext.getChildren() is a hotspot during open/close editors.
The Set semantics for EclipseContext.children are not needed since
children are anyway only added once. Only a Bag is needed.
Also it is not needed to compute a Set that is returned by getChildren.
All Callers only use the for-each loop so its sufficient to just return
an Iterable which removes stale entries on the fly instead of copying
the set.
Also children can remove from parent by just clearing their self
reference. No need to search for it.

Change-Id: If1be10ba1bebfdd2901fe96abc9a9c6e0b47c6e2
Signed-off-by: Joerg Kubitz <jkubitz-eclipse@gmx.de>
Reviewed-on: https://git.eclipse.org/r/c/platform/eclipse.platform.runtime/+/185959
Tested-by: Platform Bot <platform-bot@eclipse.org>