Bug 576350 - Add cancel user operation button to blocking dialog
This change adds a cancel button to the "User Operation Waiting"
blocking dialog, allowing the user to cancel their blocked operation in
order to continue using the Eclipse UI.

Note that the button does not cancel the operations that are blocking
the initial user operation. If the user attempts the same operation, the
blocking dialog will pop-up again (for as long as the blocking
operations are running).

Change-Id: I88284c96a7cd89133b3af55d849e3fdfcc342025
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>
Reviewed-on: https://git.eclipse.org/r/c/platform/eclipse.platform.ui/+/185969
Tested-by: Platform Bot <platform-bot@eclipse.org>
Tested-by: Lars Vogel <Lars.Vogel@vogella.com>
Reviewed-by: Lars Vogel <Lars.Vogel@vogella.com>
Reviewed-by: Andrey Loskutov <loskutov@gmx.de>