Bug 578166 - Signing information apparent in the Plug-ins tab of the About dialog
Introduce AboutBundleData.ExtendedSigningInfo as a mechanism for
allowing the user interface to be adapted to present extended signing
information, in particular p2's PGP signing information. The
AboutPluginsPage will get and adapter of this new type and the provides
it to AboutBundleData which in turn makes it available to the
BundleSigningInfo tray for presentation to the user.

Make org.eclipse.equinox.p2.ui a friend so it can provide the adapter.

Change-Id: I03de5edbebfc4858a15f0842616884df9c936396
Signed-off-by: Ed Merks <ed.merks@gmail.com>
Reviewed-on: https://git.eclipse.org/r/c/platform/eclipse.platform.ui/+/190738