Bug 578864 - remove fake job from waiting progress list
The change removes an entry for the swt display thread
"The user is waiting for background work to complete"
if no particular Job was running in display thread.

Since it is now possible to cancel exactly that job
with a dedicated cancel button the additional entry
for does not add value but is confusing.

Change-Id: I857fac569f58fb46fa0994b760b79df72a5f8da4
Signed-off-by: Joerg Kubitz <jkubitz-eclipse@gmx.de>
Reviewed-on: https://git.eclipse.org/r/c/platform/eclipse.platform.ui/+/191000
Tested-by: Platform Bot <platform-bot@eclipse.org>