Bug 579019 - Clicking on NotificationPopup should reveal parent window
NotificationPopup is always shown on the top of desktop environments
windows stack. If there are any 3rd windows on the top of the parent
window, it is not clear to what window the popped-up notification
belongs to. On clicking on the notification nothing happens. This is
confusing.

With this change: parent window is shown (brought upfront) on
clicking the NotificationPopup.

The clients that don't want the new behavior may override
addWindowActivationHelper() method to do nothing, or they might want to
extend the new behavior (do something else) by overriding
createWindowActivationHelper().

Change-Id: I1b0ebf1b5868c271b276adaa2562be36a7fc1d48
Signed-off-by: Mykola Zakharchuk <zakharchuk.vn@gmail.com>
Reviewed-on: https://git.eclipse.org/r/c/platform/eclipse.platform.ui/+/191296
Tested-by: Platform Bot <platform-bot@eclipse.org>
Reviewed-by: Andrey Loskutov <loskutov@gmx.de>