Bug 578820 - Use new FileLocator.getBundleFileLocation(Bundle)
org.eclipse.core.runtime re-exports org.eclipse.equinox.common and is
therefore one of the main suppliers of the latter plug-in. The first one
is also referenced by other plug-ins that use  the new method.
Therefore the required version has to be restricted to the first one
that provides it.

Change-Id: Idacf5b0648ceb26b6f33d277db505002aee26f89
Signed-off-by: Hannes Wellmann <wellmann.hannes1@gmx.net>
Reviewed-on: https://git.eclipse.org/r/c/platform/eclipse.platform.runtime/+/191933
Tested-by: Platform Bot <platform-bot@eclipse.org>
Reviewed-by: Andrey Loskutov <loskutov@gmx.de>