Bug 576377 - Provide shortcuts/commands for incremental multiselection/multiple carets in text editors
Added two new commands that add a caret/selection in the line
above/below and the reverse operation. This operation is similar to
what is possible with block selection, but currently treats positions
after end of line differently (currently no automatic padding to fill up
shorter lines is performed).

moveOffsetByLines now respects the visual positions instead of purely
working with document offsets, which improves usability in case of
a different number of leading tabs in adjacent lines.

Change-Id: I01563e6ab148655d0dd6c34e2a89776a1ea0f9a9
Reviewed-on: https://git.eclipse.org/r/c/platform/eclipse.platform.text/+/192202
Tested-by: Mickael Istria <mistria@redhat.com>
Reviewed-by: Mickael Istria <mistria@redhat.com>