Bug 543660 - JobGroup.join() blocks if scheduling more jobs as seed count
endJobGroup() should not unconditionally reset the
seedJobsRemainingCount to seedJobsCount. In case seedJobsCount > 1, and
we schedule only one more job, join() on the group will block forever,
waiting for the seedJobsCount to become zero.

Also jobGroup results should be preserved if new jobs are scheduled on
the group where previously scheduled jobs already finished with non-zero
results.