Wake up sleeping jobs before waiting on them
CharsetDeltaJob.addToQueue(ICharsetListenerFilter) uses timeout of 500
ms before it starts working, so it may awake in the middle of the test.

So if we want to wait for some jobs in the test, we should awake them
first, otherwise they won't be in the running/waiting state if we check
that in waitForJobs() so we will not wait for them at all.

Most likely fixes issue
https://github.com/eclipse-platform/eclipse.platform.resources/issues/42