Allow the ProgressProvider to control the used instance of a monitor
This enables the ProgressProvider to decide the instance of the
ProgressMonitor used in beginRule / join / suspend operations.

This is important if the progressprovider needs to react to certain
actions, e.g. in an UI a provider might make sure the UI is responsive
or show a "terminate" button even if the job is currently waiting on a
rule.

Fix https://github.com/eclipse-platform/eclipse.platform.runtime/issues/40