[trivial] disconnectFromWorkspace() unconditionally opens error dialog
The code in IDEWorkbenchAdvisor.disconnectFromWorkspace() relies on the
fact that ErrorDialog.openError() "will only be displayed if there is at
least one child status matching the mask".

While that is technically OK, it surprises a lot to see the code outside
of the if(status is bad) block trying to open an error dialog. If we
check for errors anyway, let open the error dialog only in that case.

Change-Id: Id7567d27b6fafa8a1d90ae978f85f23f5729fcce