Disable MarkerSortUtil batch sorting. (Fixes #53)
Can be reenabled with org.eclipse.ui.MarkerSortUtil.batchSize=10000

Batch sorting was intended to decrease memory consumption during sorting
Marker View. But it is very slow especially when CollationKeys are
calculated. Today temporary memory consumption is not as much an issue
any more, but performance is. Also test showed that the memory benefit
is not archived with batching.

Performance difference is visible with MarkerSortUtilTest especially
when increasing ARRAYSIZE to a million (22sec vs 2sec) - even without
CollationKeys.