Modernized help UI prototype: minor search and navigation improvements
Improve the modernized help UI prototype as follows:

1. Focus content frame on opening a new topic:
When opening a topic via TOC, by selecting a search result or by opening
a non search bookmark, the content frame should be focused in order to
be able to use the keyboard for scrolling (e.g. via the UP/DOWN keys).

2. When clicking on a search result, scroll to the first hit:
Opening a search result should automatically scroll to the first hit
(which worked in the past only for results of the instant search, but
not for results of a full search).

3. Opening a search bookmark should set the search field:
When opening a bookmark of a search, the search field should display the
query.

4. In Information Center, highlight search terms also when opening a
search result in a new tab:
When opening a search result in a new tab (via Ctrl+click or via
right-click and choosing "Open Link in New Tab"), the search terms
should be highlighted in the same way as when opening the result in
place.

5. In Information Center, when opening search result in a new tab fill
search field:
When opening a search result in a new tab or in a new window or when
going back in the history to a topic that was opened as a search result,
the search field should show the query. Going back in the history
applies not only for the Information Center.