Add equals() implementation for WizardCollectionElement class
It seems the TreeViewer of the ImportExportPage is refreshed when the
filter text changes.

As this text/field is selected by default when the page is shown,
clicking anywhere outside the view triggers a change event, causing the
tree to refresh and re-evaluate the collapsed/expanded state of its
children. However, this relies on the equality of its items.