NPE in ProjectNaturesPage (fixes #145)
The page is meant for elements that adapt to IProject (see plugin.xml).
However, the code adapts to IResource and casts to IProject, probably by
accident. This leads to NPE for objects that adapt to IProject, but not
to IResource.

Change to adapt to IProject directly, to have plugin.xml and code in
sync.