Improve async completion popup for completion on typing.
The popup is shown on type event of a letter or digit if the async
completion is active.
The async completion popup always shows the progress and the resulting
message of completions. The fix improves it so if async completion is
triggered automatically, then the popup is only shown if there are
completions to be shown.
A general improvement is done to close the popup when last filtered
offset is preceded by a space.