Update the static MessageDialog#openQuestion to UX guidelines from the OS
Operating system Ui guidelines do not recommend the usage of the help /
question
icon in a dialog. This change adjusts the MessageDialog#openQuestion
accordingly.As this will affect all callers of this method, it has will
lead a consistency for callers, users which want / need to violate UX
guidelines of the OS can switch to the non static methods or use
PlainMessageDialog to construct their dialog in a way which they want.

The reasoning for this change is that platform is allowed to adjust its
UI to UX changes in the OS and that the UX of Eclipse is not considered
API.

For example check the screenshots in the Windows
guidelines for dialogs.

Operating System UI Guidelines

Windows UI guidence
https://docs.microsoft.com/en-us/windows/apps/design/controls/dialogs-and-flyouts/dialogs

Gnome UI guidence
https://developer.gnome.org/hig/patterns/feedback/dialogs.html

Mac UI guidence
https://developer.apple.com/design/human-interface-guidelines/components/presentation/alerts/