WIP - Moving the perspective switcher to the left trimbar for new workspaces
Tools like VsCode, the upcoming Thunderbird, etc place their
"perspective switcher" to the left side of the application bar. With
this change we do the same for new workspaces.

This is currently a proof-of-concept, AFAIK Gayan is looking into
updating the UX of Eclipse to a more modern standard and I hope this
help him in his evaluation.