Updating UITestCase to use not extend JUnit3 anymore.
This removes the dependency of UITestCase to JUnit3. Tests which extend
UITestCase are updated to use static imports instead of inheritance to
access the assert statements.

As all tests have been upgraded to JUnit4 it is strange to still depend
on JUnit3 via inheritance.

Adds small fail wrapper in UITestCase to make migration simpler as this
requires smaller changes in the individual tests.

Not finished, as this requires more changes in the tests, e.g. the test
suite handling.