Resolve E3 views not being filtered by activities
The view dialog will always contain all available E3 views, even if some
of them have been disabled via activities.
This is because the method isFilteredByActivity uses a list of filtered
views to determine, whether the provided element should be displayed in
the dialog.
Because the element is not contained in this list, the method is unable
to detect filtered activities and therefore always returns false.
An additional check ensures that this filtering isn't done to E4 views,
as they are also not contained in the view registry.
Foo