Make E4 more accessible in org.eclipse.ui to ease migrations
Currently it is very hard to migrate parts of the code, there is only
all or nothing and migration feels stuck for years now.

One culprit is the exhaustive use of PlatformUI.getWorkBench() and the
inability for code to easily "join" the E4 ecosystem without major
rewrites.

This change proposes a new API in PlatformUI that allows a migration
path from old Workbench API to E4 without creating an immediate
requirement to refactor all components. As an example
ObjectContributorManager is refactored to use the new API so it could
access the IExtensionTracker in both worlds.