Make WizardDialog standard PrimaryModal not ApplicationModal
The wizard dialog should not block all windows that are opened in
eclipse, but just the window that did trigger it, which is its parent.
Because very likely the other window has not much todo with the context
of this wizard dialog, which is then for users annoying that it will
block.
With PrimaryModal it will only block the parent window(s)

fixes #242