Bug #251 - ShowViewMenu does not work for E4 Views
It looks like the org.eclipse.ui.internal.ShowViewMenu.fillMenu(IMenuManager) method is only looking in the E3 View list for the view id: WorkbenchPlugin.getDefault().getViewRegistry();
Whereas the Dialog looks in the E4 Descriptor List: ((MApplication) element).getDescriptors();

E4 Views are now visible in the Window > Show View Menu and the Customize Perspective dialog.

How To Test:
Add a E4 Part Descriptor as a View Shortcut in a Perspective ether in code or by using the Customize Perspective Dialog (which is now capable of showing E4 Parts), the E4 Part should now be visible in the Window > Show View quick access menu.