Bug 578859 - [MinMaxAddon] NPE on maximize with sub Area (#257)
MinMaxAddon checks now if the found Area has a SharedRef, if this is not the case it searches for the next higher Area.

How to test:
Create inside the Editorss Area a Composite Part containing an Area which contains multiple parts (see bug ticket for example). Double click one of the child parts, the editorss area should now be maximized.