Remove org.apache.commons.logging from o.e.e4.rcp feature
If a Feature includes a Plugin, it usually includes it with a specific
name and the version that was in the TP when the feature was build.
This prevents consumers from using 'logging-bridges' in their TP or
product, for example a slf4j-bridge that provides binary compatible
packages and classes but redirects all events to slf4j.