Include org.eclipse.e4.ui.progress in org.eclipse.e4.rcp
The org.eclipse.equinox.p2.ui bundle requires the
org.eclipse.e4.ui.progress package of the org.eclipse.e4.ui.progress
bundle so the bundle is already installed with most applications.

Furthermore, if this bundle is not included in any feature, it will need
special case handling for publishing to Maven, which would be best
avoided.

https://github.com/eclipse-platform/eclipse.platform.releng/issues/126