Show editor guiding text #78
Adds an "OnboardingComposite" to editor CTabFolder.
The OnboardingComposite contains a greyed image of the current
perspective and a "Welcome to <perspective" text.

A paintlistener controls the size of this OnboardingComposite.
If no editor is open, the OnboardingComposite is shown. As soon as
editor opens it is hidden.

A perspective switch listener changes the image and the label text.

On another change this will be further implemented, e.g. by a new
extension point. In this extension point a larger image could be
defined. And also some more text, e.g. a bullet list with action ("What
can be done in this perspective?").

Signed-off-by: Marcus Hoepfner <marcus.hoepfner@sap.com>