Project wizards should be shown in "new project" category
Currently wizard that are marked with project=true are shown with other
generic ones instead of in the "new project" section of the navigator.

This adds them to the "new project" category so they are easier
identified by a user that they create a new project.

Fix https://github.com/eclipse-platform/eclipse.platform.ui/issues/502