Revert "Revert "Remove explicit requirement to jxpath in feature""
This reverts commit c4135cf4ee86c08b85e71314fa0f62d3e1012e89.
The former commit was reverted because of a bug in Tycho and because
the build was still consuming jxpath from Maven instead of Orbit.
Now that 1 build succeeded, there should be nothing left of the jxpath
from Maven Central, and removing this transtive dependency from the
feature should simplify further maintenance and evolution.