Ctrl-E: Do not handle name collisions in case of custom editor input
In case the editor input returns a 'name' that is not the actual file
name, do not apply the collision resolving logic based on file path
segments.

In such cases we need to assume that the editor input is custom-built
and such files might be virtual, in which case showing a path is not the
intended behavior.

Fixes #547.