Show editor guiding text #78 (#78)
Adds an "OnboardingComposite" to editor CTabFolder.
The OnboardingComposite contains a greyed image of the current
perspective and a "Welcome to <perspective" text.

A paintlistener controls the size of this OnboardingComposite.
Currently it always hides (commented code) until the extension point is
enhanced/added. See below.

A perspective switch listener changes the image and the label text.

On another change this will be further implemented, e.g. by a new
extension point. In this extension point a larger image could be
defined. And also some more text, e.g. a bullet list short cuts.

Signed-off-by: Marcus Hoepfner <marcus.hoepfner@sap.com>