Check preconditions in FileSystemResourceManager.write #103
The FileSystemResourceManager.write() method fails when the target file
has been deleted before the write method has been called. The method
defines the existence of the file in the workspace as a precondition but
does never check that, which makes the method fail when first accessing
information that is not available after the deletion, in particular, its
ResourceInfo. This is why the method failed when accessing the
ResourceInfo.

The FileSystemResourceManagerTest.testWriteFile() randomly fails because
of this behavior. A concurrently running workspace refresh removes the
file right before the write() method is called.

- Make write() check its precondition, i.e., the existence of the target
file to deterministically fail with an IllegalStateException
- Make testWriteFile() wait for the refresh to be finished before
executing write() to have deterministic test behavior
- Add additional test case validating proper failure of write() when
target file has been deleted before

Fixes https://github.com/eclipse-platform/eclipse.platform/issues/103