EclipseContext: allow GC for undisposed contexts #375
* WeakGroupedListenerList:
Use JDK's WeakHashMap (with a cleaner) to not leak
WeakComputationReference

* InjectorImpl:
Use WeakHashMap to avoid strong references to injected Contexts to avoid
leaked EclipseContext in the key of InjectorImpl.injectedObjects

* Fix modifiers and their order

Fixes Eclipse-IDE process stuck in EclipseContext.cleanup during
shutdown. Happened if EclipseContext could be garbage collected but is
not explicitly disposed.