Make ResourceTest restore workspace description #406
Several realizations of a ResourceTest perform changes to the workspace
description, e.g., in terms of changing the auto-build settings, without
resetting them afterwards. And even if they do, the functionality is
redundantly present at different places in source code.
This induces test dependencies, as the behavior of a subsequent test
case execution depends on the modifications of the workspace description
of previous ones.

- Store the workspace description before executing a ResourceTest and
  restore it during teardown
- Provide only one central method enabling and disabling autobuild and
  setting the build order and remove all redundant methods or in-place
  realizations
- Remove manual restorations of autobuild settings after test execution,
  as this is subsumed by ResourceTest setup and teardown
- Properly wait for background jobs completion in BuilderTest to avoid
  race conditions because of changed runtime behavior due to resetting
  workspace description after executing other test cases