Make sure the title image is always updated in CompatibilityPart
When updating the image from the workbench part currently a trivial put
operation is used, but because of how the transient data map works it
only triggers a change if the value is not equal to the previous one.

It can now happen that two images are considered equal because they are
created from the same imagedescriptor, but one of them is disposed
afterwards in which case the value in the transient data is stale even
though it was tried to being updated.

This is now changed in the following:
- first the current value is compared with the update value using
identity comparison
- if it is not the same object, a remove/put operation is performed to
enforce an update of the value in the map

Fix https://github.com/eclipse-platform/eclipse.platform.ui/pull/682