Wizard pages should be able to specify their minimum size.
Adds the new methods getMinimumWizardSize() to the IWizard and
getMinimumPageSize() to the IWizardPage interface. It is currently not
possible to control the minimum size of the wizard page, if the wizard
is not created by the application. You can take the target platform
editor as an example, where custom wizards may be contributed via an
extension point.

Subclasses of WizardPage can overwrite the getMinimumPageSize() method
in order to set a minimal size for the wizard shell. By default, no such
minimum size is set.

If getMinimumPageSize() is not overwritten, then the minimum size of the
shell is not set. This corresponds to the current behavior. Otherwise
the minimum size is of the shell is calculated to be equal or larger
than the minimum sizes of all wizard pages.